Move connect-to-mongodb.js to src/modules/
