+++ src/main.js
@@ -4,7 +4,7 @@ require('dotenv').config();
 const express = require('express');
 const { graphqlHTTP } = require('express-graphql');
 
-const connectToMongoDb = require('./connect-to-mongodb');
+const connectToMongoDb = require('./modules/connect-to-mongodb');
 const graphqlSchema = require('./schemas/graphql-schema');
 
 const { PORT } = process.env;
